issue #218: refactor the startup behaviour
