+++ code/service/src/main/resources/application.conf
@@ -42,6 +42,7 @@ akka {
 		journal.plugin = "jdbc-journal"
 		snapshot-store.plugin = "jdbc-snapshot-store"
 	}
+
 	cluster {
 		sharding {
 			# Number of shards used by the default HashCodeMessageExtractor
@@ -56,7 +57,6 @@ akka {
 			# It is always disabled if `remember-entities` is enabled.
 			passivate-idle-entity-after = 120s
 		}
-		downing-provider-class = "akka.cluster.sbr.SplitBrainResolverProvider"
 		# Documentation: https://doc.akka.io/docs/akka/current/split-brain-resolver.html#split-brain-resolver
 		downing-provider-class = "akka.cluster.sbr.SplitBrainResolverProvider"
 		split-brain-resolver {
@@ -88,10 +88,12 @@ akka {
 		}
 		shutdown-after-unsuccessful-join-seed-nodes = 20s
 	}
+
 	coordinated-shutdown {
 		terminate-actor-system = on
 		exit-jvm = on
 	}
+
 	projection {
 		slick {
 			# The Slick profile to use
